This implement the load part of OFBIZ-734 (POS Save and load sessions). Actually they are not sessions but sales. You can add a shopping list (saved sale) to the current cart (current sale) or replace it. The button to use is "LOAD SALE" (or "Charger une vente" in french) This work was partly funded by Ray Barlow, thanks to him.
In the same time decided to commit the xpos.properties file in the state it is on machine to avoid futur svn conflicts for me. If someone thinks it's a problem let me know.
I commit also additions in posmain.xml : new product button to test the new scrollable panel.

git-svn-id: https://svn.apache.org/repos/asf/incubator/ofbiz/trunk@428156 13f79535-47bb-0310-9956-ffa450edef68