Patch from JIRA issue https://issues.apache.org/jira/browse/OFBIZ-1910 "Add support of fromDate & thruDate in CustRequestRole entity".
Slight change in Entity name in this commit.
In the attached patch the  name of entity was "CustomerRequestRole" but in this commit the name of new entity is "CustRequestParty"(Looked more close to the data model pattern).

Thanks Bilgin for your valuable Input on this.
I have also written the Migration service for providing the Backward compatibility.

Special thanks to Ratnesh & Jacopo in discussing with me on this work.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@686731 13f79535-47bb-0310-9956-ffa450edef68