VisualThemes for Ecommerce (https://issues.apache.org/jira/browse/OFBIZ-2106)
The ProductStore entity has now a visualThemeId that selects the VisualTheme to be used in the Ecommerce.
This can be set in the EditProductStore form selecting a theme between the themes that belong to the "ECOMMERCE" VisualThemeSet

To the following ProductStore fields has been applied the entity deprecation pattern:
<field name="oldStyleSheet" col-name="STYLE_SHEET" type="url"></field>
<field name="oldHeaderLogo" col-name="HEADER_LOGO" type="url"></field>
<field name="oldHeaderMiddleBackground" col-name="HEADER_MIDDLE_BACKGROUND" type="url"></field>
<field name="oldHeaderRightBackground" col-name="HEADER_RIGHT_BACKGROUND" type="url"></field>

The first part of the ecommerce header.ftl (the part defining the <HEAD> tag content) has been moved to a new headerHead.ftl file. This allows all the theme defined header.ftl to include this common part that is mainly inclusion of theme defined resources and should never change (code reuse).

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@746544 13f79535-47bb-0310-9956-ffa450edef68