A patch from Angelo Matarazzo "Add view-last-noparam" (https://issues.apache.org/jira/browse/OFBIZ-3027) - OFBIZ-3027 "view-last-noparam" has functionality of view-last but without last parameters like "request-redirect-noparam".
I' m using this attribute because in my application all the screens are linked to a field in the context. that I call "commonValue". I have a link in the common header that change the value of this field but when I click this link the screen must not change but the field of current screen must be cleaned. By this attribute the view is set at " last-view" but the last parameters are cleaned.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@829727 13f79535-47bb-0310-9956-ffa450edef68