A suggested change by Patrick Antivackis "Shopping Cart Item issue with BigDecimal " (https://issues.apache.org/jira/browse/OFBIZ-4038) - OFBIZ-4038
In applications/order/src/org/ofbiz/order/shoppingcart/ShoppingCartItem.java
the setQuantity method should use this.quantity.compareTo(quantity) == 0 instead of this.quantity==quantity like depicted below

   protected void setQuantity(BigDecimal quantity, LocalDispatcher dispatcher, ShoppingCart cart, boolean triggerExternalOps, boolean resetShipGroup, boolean updateProductPrice, boolean skipInventoryChecks) throws CartItemModifyException {
       if (this.quantity == quantity) {            return;        }

The issue is really visible when the cart is fully loaded (lot of products) as the return will never occur and the store is configurerd in an autosavecart mode.

For example a cart with 48 products will take very long time to process for each additem or modifycart event.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1042188 13f79535-47bb-0310-9956-ffa450edef68