Some work on ContainerLoader.java - Fixes some thread-safe issues and try to have better control over loading/starting/stopping containers.
I'm not sure what the Geronimo code was supposed to do. It appeared to load all of the containers just so it could create/start the rmi-dispatcher container. So, I created a static method that allows client code to get a specific container. If this change causes problems with Geronimo just let me know and I will fix it.


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1084238 13f79535-47bb-0310-9956-ffa450edef68