Closes "Login out on a cluster handled by DeltaManager causes a NPE" https://issues.apache.org/jira/browse/OFBIZ-4289
It works locally but not on clusters. You simply get a NPE.

Problem
    When we logout we cross an issue due to Tomcat sessions persistence.
    Because we set distributable to true, to allow sessions failover,
    and use and DeltaManager for replication.
    Delegator and other main Classes (notably Dispatcher) are not serialized in OFBiz. This is the origin of the problem

Tried solutions
By default DeltaManager save active sessions on disk. It uses a private String variable pathname for that (default to "SESSIONS.ser"). You can set it to null, to avoid session persistence, using a context.xml file in the WEB-INF folder with content like below. But I tried it in webtools app on staging qs001 (only) and it did not work (with distributable set to false). And we have no other means to set it from OFBiz (OOTB).

<Context>
     <Manager className="org.apache.catalina.session.DeltatManager" pathname="">
     </Manager>
</Context>



git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1132749 13f79535-47bb-0310-9956-ffa450edef68