A patch from Pierre Smits "Project resource overview doesn't show the resources with a thru-date before the current date." https://issues.apache.org/jira/browse/OFBIZ-4853
When a project manager expires the assignment of a resource by clicking the delete button in the Project Resources overview the resource is deleted from the overview. It should be that the resource is still shown in the overview with the tru-date visible.
Otherwise there will be a discrepancy between the overview of resources of the project and the project dashboard.
The name of the button is also wrong. It should state 'expire' not 'delete'.

jleroux: I have also created a new label FormFieldTitle_expireButton copying labels from CommonExpire. I saw that labels where not sorted at all so I used Labels Manager for that. I put back CommonEmptyHeaderby hand. I will soon fix this issue. So, apart the sort changes, there is also a lot of false changes in CommonUiLabels.xml because Labels Manager now uses the right way for Language Identification: http://www.w3.org/TR/REC-xml/#sec-lang-tag (ie a - separate language from country instead of previously a _)

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1338695 13f79535-47bb-0310-9956-ffa450edef68