A modified patch from Karl Beecher for "Updating EFT Account address resurrects out-of-date EFT Accounts" https://issues.apache.org/jira/browse/OFBIZ-5193
When you update the address that an EFT Acccount is linked to, all expired EFT accounts are "resurrected", which can mean multiple out-of-date copies of an EFT account are displayed in the Payment Method Information.

To reproduce (with OFBiz demo data):

* Go to DemoSupplier profile page and click Create EFT Account
* Enter details of a new account, making sure the account is linked to the DemoSupplier postal address.
* Click Save and then click Cancel/Done.
* Back on the profile page, Update for this new EFT account.
* Change at least one field of the account.
* Click Save (now there are two EFT account records in the database; one of which is expired). Click Cancel/Done.
* Now click Update for the DemoSupplier's postal address.
* Update at least one field. Click Save.
* Go back to the DemoSupplier profile page. You should see that the old expired EFT accounts are listed in the Payment Method Information.

We avoid this bug by altering PaymentMethodServices.xml#setPaymentMethodAddress, so that out-of-date EFT Accounts are filtered out.

jleroux: I applied changes by hand because the patch did not work. I also removed the useless line setting a now var.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1527254 13f79535-47bb-0310-9956-ffa450edef68